+++ codebase/apps/physics/src/ModelDerive/DeriveBase.cc
@@ -68,7 +68,7 @@ DeriveBase::~DeriveBase()
 
 int DeriveBase::derive(vector<float*> *inputs, vector<float*> *outputs) 
 {
-  if(inputs->size() != this->numberOfInputs) {
+  if (inputs->size() != static_cast<size_t>(this->numberOfInputs)) {
     return -2;
   }
   return 0;
