+++ codebase/apps/physics/src/ModelDerive/ModelDeriveMdv.cc
@@ -164,7 +164,7 @@ bool ModelDeriveMdv::processData(TriggerInfo &trigger_info)
 
     //
     // Save the output fields 
-    for(int d = 0; d < output_data.size(); d++) {
+    for (size_t d = 0; d < output_data.size(); d++) {
       //
       // If the output field units are "inputs" set its units to the inputs units
       if(strcasecmp((*units)[d].c_str(), "inputs") == 0) {
@@ -186,7 +186,7 @@ bool ModelDeriveMdv::processData(TriggerInfo &trigger_info)
   //
   // Copy over requested fields
   //
-  for(int a = 0; a < fields_to_copy.size(); a++) {
+  for (size_t a = 0; a < fields_to_copy.size(); a++) {
 
       MdvxField *field = input_mdv.getField(fields_to_copy[a]);
       if(field == 0)
@@ -318,7 +318,7 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
 	continue;
 
       int field_index = -1;
-      for(int d = 0; d < fhdrsFile.size(); d++) {
+      for (size_t d = 0; d < fhdrsFile.size(); d++) {
 	if( strcasecmp( (fhdrsFile[d]).field_name_long, _params->__derive_functions[a][b]) == 0 ||
 	    strcasecmp( (fhdrsFile[d]).field_name, _params->__derive_functions[a][b]) == 0) {
 	  field_index = d;
@@ -329,7 +329,7 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
       if (field_index == -1) 
       {
 	bool found = false;
-	for(int c = 0; c < deriveFields.size(); c++) {
+	for (size_t c = 0; c < deriveFields.size(); c++) {
 	  if(strcasecmp(deriveFields[c].c_str(), _params->__derive_functions[a][b]) == 0) {
 	    found = true;
 	    break;
@@ -348,11 +348,11 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
       
     }
     const vector<string> *shortNames = DeriveBase::getOutputShortNames();
-    for(int d = 0; d < shortNames->size(); d++)
+    for (size_t d = 0; d < shortNames->size(); d++)
       deriveFields.push_back( (*shortNames)[d] );
 
     const vector<string> *longNames = DeriveBase::getOutputLongNames();
-    for(int d = 0; d < longNames->size(); d++)
+    for (size_t d = 0; d < longNames->size(); d++)
       deriveFields.push_back( (*longNames)[d] );
 
   }
@@ -368,7 +368,7 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
 	continue;
 
       bool found = false;
-      for(int c = 0; c < deriveFields.size(); c++) {
+      for (size_t c = 0; c < deriveFields.size(); c++) {
 	if(strcasecmp(deriveFields[c].c_str(), _params->_interpolate_function[b]) == 0) {
 	  found = true;
 	  break;
@@ -376,7 +376,7 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
       }
       if(found == false) {
 	int field_index = -1;
-	for(int d = 0; d < fhdrsFile.size(); d++) {
+	for (size_t d = 0; d < fhdrsFile.size(); d++) {
 	  if( strcasecmp( (fhdrsFile[d]).field_name_long, _params->_interpolate_function[b]) == 0 ||
 	      strcasecmp( (fhdrsFile[d]).field_name, _params->_interpolate_function[b]) == 0) {
 	    field_index = d;
@@ -411,7 +411,7 @@ int ModelDeriveMdv::_create_verify_input_field_list(DsMdvx &input_mdv)
 	continue;
 
       int field_index = -1;
-      for(int d = 0; d < fhdrsFile.size(); d++) {
+      for (size_t d = 0; d < fhdrsFile.size(); d++) {
 	if( strcasecmp( (fhdrsFile[d]).field_name_long, _params->_copy_fields[b]) == 0 ||
 	    strcasecmp( (fhdrsFile[d]).field_name, _params->_copy_fields[b]) == 0) {
 	  field_index = d;
